# engine/openapi.yaml
openapi: 3.0.3
info:
  title: OpenPipe Workflow API
  version: 0.7.0
  description: API for discovering, validating, diagnosing, syncing, scaffolding, and running OpenPipe workflows/pipes.
servers:
  - url: http://127.0.0.1:8787
paths:
...
  /runs:
    get:
      summary: List recent run reports
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Recent runs
  /runs/summary:
    get:
      summary: Summarize recent run outcomes
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Run summary
  /runs/trends:
    get:
      summary: Run trends grouped by UTC day
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Run trends
  /workflows/scaffold:
    post:
      summary: Scaffold a new workflow pipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [packId]
              properties:
                packId:
                  type: string
                baseDir:
                  type: string
      responses:
        '200':
          description: Scaffold completed
        '409':
          description: Pipe already exists
  /workflows:
    get:
      summary: List workflows
      parameters:
        - in: query
          name: details
          required: false
          schema:
            type: boolean
          description: Include workflow metadata summaries when true.
      responses:
        '200':
          description: Workflow list
...
  /pipes:
    get:
      summary: Alias of /workflows
      parameters:
        - in: query
          name: details
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Pipe list
...